third_party/blink/web_tests/fast/events/bogus-event-listener-invalidation-expected.txt

This page tests for a bogus invalidation of a node's event listeners. If the test passes, you'll see a PASS message below.

PASS: div.onclick should be function () { return "click"; } and is.
